Ability to pass `format` parameter
With this it is possible to force the Adapter to save the incoming file in a specific format. This makes sense if you want to upload svg files. These are actually xml files and wont get recognized by cloudinary as images if it is not being told to do so.